fix: prevent job termination on Slurm node lookup failures
Presently, when the operator cannot query a Slurm node due to the
RestAPI being unavailble (or other reasons), the pod within which that
node is run will be killed, and any runing jobs within it terminated.
This commit changes this behavior such that SlurmClient does not check
whether the RestAPI is ready when running its' reconciler loop. This
should prevent the premature termination of NodeSet pods due to the
RestAPI being unavailable.

Changelog: fixed - Modified SlurmClient behavior to prevent job
 termination on Slurm job lookup failures