fix: fixes e2e-test execution by introducing pre-test cleanup
Previously the e2e-test would not consistently pass. This was due to an
issue with the retry logic that was implemented for the srun test.
Occasionally, when the first srun was launched prior to the controller
being fully ready, the job would not be fully cleaned up be the next
iteration of the retry loop. This would result in the second srun
getting stuck pending, causing the tests to hang and fail. This fix
enforces the exection of scancel prior to the execution of each srun,
ensuring a cleaner test environment.